+++ Changelog.md
@@ -1,5 +1,6 @@
 ### 0.2.3
 + Add slack mode
++ Implement Reformation for char
 
 ### 0.2.2
 + Auto replacement of capturing groups with non-capturing