+++ Changelog.md
@@ -1,7 +1,8 @@
 ### 0.2.3
 + Add slack mode
 + Implement Reformation for char
-+ Closing on enum support
++ Enum support
++ Tuple struct support
 
 ### 0.2.2
 + Auto replacement of capturing groups with non-capturing